+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 ### Changed
 
-- Go version updated from `1.16.3` up to `1.16.6`
+- Go version updated from `1.16.3` up to `1.16.7`
 
 ## v4.3.0