+++ CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## UNRELEASED
+
+### Changed
+
+- Go version updated from `1.17.2` up to `1.17.3`
+
 ## v4.4.0
 
 ### Changed
 
-- Go version updated from `1.16.3` up to `1.17.3`
+- Go version updated from `1.16.3` up to `1.17.2`
 
 ### Added