+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## v4.5.1
+
+### Changed
+
+- Go dependencies updated
+
 ## v4.5.0
 
 ### Changed